+++ CookingSkill/docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## 1.3.1
 * Fixed multiplied recipes in Cooking SKill 1.3.0.
+* Internal refactoring.
 
 ## 1.3.0
 Released 10 July 2021 for Stardew Valley 1.5. Updated by Pathoschild.